+++ app_dart/lib/src/request_handlers/flaky_handler_utils.dart
@@ -22,8 +22,6 @@ const String kP4Label = 'P4';
 const String kP5Label = 'P5';
 const String kP6Label = 'P6';
 const String kBigQueryProjectId = 'flutter-dashboard';
-const String kCiYamlPath = '.ci.yaml';
-const String kTestOwnerPath = 'TESTOWNERS';
 const String kCiYamlTargetsKey = 'targets';
 const String kCiYamlTargetBuilderKey = 'builder';
 const String kCiYamlTargetIsFlakyKey = 'bringup';
